requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
DateTime==4.3
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
mysqlclient==1.4.2
pycparser==2.19
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
zope.interface==4.6.0
